--- src/pages/strategies2/index.js
@@ -1,132 +1,0 @@
-
-
-// ** React Imports
-import { useState, useContext } from 'react'
-
-// ** MUI Imports
-import {Box } from '@mui/material'
-import toast from 'react-hot-toast'
-
-// ** Custom Imports
-import Table from 'src/components/Shared/Table'
-import GridToolbar from 'src/components/Shared/GridToolbar'
-
-// ** API
-import { RequestsContext } from 'src/providers/RequestsContext'
-
-import { FinancialRepository } from 'src/repositories/FinancialRepository'
-
-// ** Windows
-
-// ** Helpers
-import ErrorWindow from 'src/components/Shared/ErrorWindow'
-import { useInvalidate, useResourceQuery } from 'src/hooks/resource'
-
-// ** Resources
-import { ResourceIds } from 'src/resources/ResourceIds'
-import { DocumentReleaseRepository } from 'src/repositories/DocumentReleaseRepository'
-
-const Strategies2 = () => {
-  const { getRequest, postRequest } = useContext(RequestsContext)
- 
-  const [selectedRecordId, setSelectedRecordId] = useState(null)
-
-  //states
-  const [windowOpen, setWindowOpen] = useState(false)
-  const [errorMessage, setErrorMessage] = useState(null)
-
-  async function fetchGridData(options = {}) {
-    const { _startAt = 0, _pageSize = 50 } = options
-
-    return await getRequest({
-        extension: DocumentReleaseRepository.Strategy.qry,
- parameters: `_startAt=${_startAt}&_pageSize=${_pageSize}&filter=`
-    })
-  }
-
-
-  const {
-    query: { data },
-    labels: _labels,
-    paginationParameters,
-    refetch,
-    access
-  } = useResourceQuery({
-    queryFn: fetchGridData,
-    endpointId:  DocumentReleaseRepository.Strategy.qry,
-    datasetId: ResourceIds.Strategies
-  })
-
-  const invalidate = useInvalidate({
-    endpointId:  DocumentReleaseRepository.Strategy.qry
-  })
-
-  const columns = [
-    {
-      field: 'name',
-      headerName: _labels.name,
-      flex: 1
-    },
-    {
-      field: 'groupName',
-      headerName: _labels.groupstra,
-      flex: 1
-    }
-  ]
-
-
-  const add = () => {
-    setWindowOpen(true)
-  }
-
-  const edit = obj => {
-    setSelectedRecordId(obj.recordId)
-    setWindowOpen(true)
-  }
-
-  const del = async obj => {
-    await postRequest({
-      extension: FinancialRepository.Group.del,
-      record: JSON.stringify(obj)
-    })
-    invalidate()
-    toast.success('Record Deleted Successfully')
-  }
-  
-
-  return (
-    <>
-      <Box>
-        <GridToolbar onAdd={add} maxAccess={access} />
-        <Table
-          columns={columns}
-          gridData={data}
-          rowId={['recordId']}
-          onEdit={edit}
-          onDelete={del}
-          isLoading={false}
-          pageSize={50}
-          refetch={refetch}
-          paginationParameters={paginationParameters}
-          paginationType='api'
-          maxAccess={access}
-        />
-      </Box>
-      {/* {windowOpen && (
-        <AccountGroupsWindow
-          onClose={() => {
-            setWindowOpen(false)
-            setSelectedRecordId(null)
-          }}
-          labels={_labels}
-          maxAccess={access}
-          recordId={selectedRecordId}
-          setSelectedRecordId={setSelectedRecordId}
-        />
-      )}
-      <ErrorWindow open={errorMessage} onClose={() => setErrorMessage(null)} message={errorMessage} /> */}
-    </>
-  )
-}
-
-export default Strategies2
